example 2 wip

+++ Example2-Survival/Assets/Svelto/Svelto.Common/DataStructures/ThreadSafeQueue.cs
@@ -112,7 +112,7 @@ namespace Svelto.DataStructures
             {
                 int i = list.Count;
                 
-                list.GrowAndSetCount(m_Queue.Count);
+                list.GrowAndSetCount((uint) m_Queue.Count);
 
                 var array = list.ToArrayFast();
